requirements.txt
fastapi
uvicorn[standard]
celery[redis]
redis
openai
python-dotenv
pydantic
transformers 
torch 